api: Split service endpoints into their own files
